% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencetable.R
\name{adjust_rt}
\alias{adjust_rt}
\title{Adjust Retention Time}
\usage{
adjust_rt(rttable, adjustColumn, adjustAnalyte, adjustMax = TRUE,
  adjustment)
}
\arguments{
\item{rttable}{the table of retention time values}

\item{adjustColumn}{which ColID you want to make the adjustment on (numeric)}

\item{adjustAnalyte}{which analyte you want to make the adjustment on (string)}

\item{adjustMax}{Whether you want to adjust the MaxRT or the MinRT (True/False)}

\item{adjustment}{Size of the adjustment, adds in both cases, so to lower the
minimum RT put a negative value in.}
}
\value{
calls the check_rt_table function to verify that the table will not
cause other problems, and then returns either the error message or the adjusted
table
}
\description{
This function can be used to adjust an individual retention time value
}
