% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakassignment.R
\name{assign_analytes}
\alias{assign_analytes}
\title{Assign Analytes}
\usage{
assign_analytes(peaktable, rttable)
}
\arguments{
\item{peaktable}{the table of peak values}

\item{rttable}{the table of retention time values}
}
\value{
Returns the analyte (as a factor) or an NA value
}
\description{
Works through a join (on ColID), so scaling may be an issue with large datasets
}
