% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencetable.R
\name{create_min_max_rt}
\alias{create_min_max_rt}
\title{Creating Minimum and Maximum RT windows}
\usage{
create_min_max_rt(df, window = 0.022)
}
\arguments{
\item{df}{data frame of retention times}

\item{window}{size of RT window, will be added and subtracted from median}
}
\value{
new reference table, will not be checkd so that a value is retuned
}
\description{
Creating Minimum and Maximum RT windows
}
