% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencetable.R
\name{find_median_rt}
\alias{find_median_rt}
\title{Find median retention times from standards}
\usage{
find_median_rt(df, analytelist = c("AA", "PA", "BA", "IS"),
  excludedpoints = NA, areacutoff = 40, frontrtcutoff = 2.2,
  rearrtcutoff = 1.8)
}
\arguments{
\item{df}{dataframe that contains all peaks from an SCFA GC run}

\item{analytelist}{a list of analytes that are in the standards, must be entered in order}

\item{excludedpoints}{a list of points that should be excluded}

\item{frontrtcutoff}{Points with retentions times less than this number will be excluded from the front column}

\item{rearrtcutoff}{Points with retention times less than thsi number will be excluded from the rear column}

\item{areacuttoff}{All points with areas less than this number will be excluded}
}
\value{
a dataframe with each GC column having all analytes assigned
}
\description{
Find median retention times from standards
}
