% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencetable.R
\name{generate_rt_table}
\alias{generate_rt_table}
\title{Generate Retention Time Table}
\usage{
generate_rt_table(df, RTwindow = 0.022, analytelist = c("AA", "PA",
  "BA", "IS"), excludedpoints = NA, areacutoff = 40,
  frontrtcutoff = 2.2, rearrtcutoff = 1.8)
}
\arguments{
\item{df}{Raw data, each peak gets its own row}

\item{RTwindow}{the +/- from the median RT for every analyte that will be adjusted}

\item{analytelist}{list of analytes that are present in the controls}

\item{excludedpoints}{list of points (untested as of 8/21)}

\item{areacutoff}{will exclude all peaks below this area count}

\item{frontrtcutoff}{cuts off all peaks before a certain retention time on the front (even) GC column}

\item{rearrtcutoff}{cuts off al peaks before a certain retention time on the rear (odd) GC column}
}
\value{
returns the retention time table
}
\description{
This function takes raw data generated by the python script
and creates a retention time table by calling a few utility
functions. It is meant to be somewhat adjustable, and other
functions can be used to modify the table.
}
\seealso{
find_median_rt

create_min_max_rts
}
