% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencetable.R
\name{sample_label}
\alias{sample_label}
\title{Pull GC related Metadata}
\usage{
sample_label(df)
}
\arguments{
\item{df}{A dataframe containing SCFA data extracted with at least the following
columns: SID, ColID}
}
\value{
newfile a dataframe that has had metadata added to it
}
\description{
`samplelabel` is a convenience function that pulls out meta-data from the
extracted datafile. Notably it splits up the SID column into component columns
of Experiment ID (EID), plate identity (PID), and Well
}
\details{
Based on how the sample name is formated it will also determine if the sample
is one of blank/standard/calibration/sample and give the appropriate label
}
